the changed experience model

+++ DeepQ_StarCraft2/train_defeat_zerglings.py
@@ -1,3 +1,4 @@
+# people experience  algorithm
 import sys
 import os
 import datetime
@@ -15,6 +16,7 @@ _SELECT_ALL = [0]
 _NOT_QUEUED = [0]
 
 FLAGS = flags.FLAGS
+# people experience  algorithm
 flags.DEFINE_string("algorithm", "deepq", "RL algorithm to use.")
 flags.DEFINE_string("log", "tensorboard", "logging type(stdout, tensorboard)")
 
@@ -23,18 +25,19 @@ flags.DEFINE_boolean("prioritized", True, "prioritized_replay")
 flags.DEFINE_bool("visualize", True, "if you want to see the game")
 
 flags.DEFINE_float("exploration_final_eps",  0.01, "your final Exploration Fraction")
-flags.DEFINE_float("exploration_fraction",  0.75, "Exploration Fraction")
+flags.DEFINE_float("exploration_fraction",  0.47, "Exploration Fraction")
 flags.DEFINE_float("gamma", 0.99, " the speed of exploration")
-flags.DEFINE_float("lr",  0.0001, "Learning rate")
+flags.DEFINE_float("lr",  0.001, "Learning rate")
 
-flags.DEFINE_integer("train_freq", 5, "the freq that you train your model")
+flags.DEFINE_integer("train_freq", 100, "the freq that you train your model")
+flags.DEFINE_integer("batch_size", 1500, "the number of your examples that you want to train your model")
 flags.DEFINE_integer("print_freq", 15, "the freq that you print you result")
-flags.DEFINE_integer("learning_starts", 10000, "Learning start time")
-flags.DEFINE_integer("timesteps", 1000000, "most Steps to train")
+flags.DEFINE_integer("learning_starts", 150000, "Learning start time")
+flags.DEFINE_integer("timesteps", 2500000, "most Steps to train")
 flags.DEFINE_integer("num_actions", 4, "numbers of your action")    #3
-flags.DEFINE_integer("step_mul", 1, "the time of every step spends")
-flags.DEFINE_integer("episode_steps", 2600, "the steps of every episode spends")# 2000
-flags.DEFINE_integer("buffer_size", 20000, "the number of actions that you want to store")
+flags.DEFINE_integer("step_mul", 5, "the time of every step spends")
+flags.DEFINE_integer("episode_steps", 2000, "the steps of every episode spends")# 2000
+flags.DEFINE_integer("buffer_size", 22000, "the number of actions that you want to store")
 flags.DEFINE_integer("target_network_update_freq", 100, "the freq that your network update")
 
 PROJ_DIR = os.path.dirname(os.path.abspath(__file__))
@@ -79,7 +82,7 @@ def main():
       game_steps_per_episode= FLAGS.episode_steps) as env:
 
     model = deepq.models.cnn_to_mlp(
-      convs=[(32, 8, 4), (64, 4, 2), (64, 3, 1)],
+      convs=[(32, 8, 4), (64, 4, 2), (64, 3, 1), (64, 3, 1), (64, 3, 1), (32, 3, 1)],
       hiddens=[256],
       dueling=True
     )
@@ -104,16 +107,16 @@ def main():
     act.save("defeat_zerglings.pkl")
 
 def deepq_callback(locals, globals):
-  global max_mean_reward, last_filename
+  global max_mean_reward, last_filename, best_reward_episode
   if('done' in locals and locals['done'] == True):
 
     print("mean_100ep_reward : %s max_mean_reward : %s" %
           (locals['mean_100ep_reward'], max_mean_reward))
 
     if ('mean_100ep_reward' in locals
-            and locals['num_episodes'] >= 500
-            and locals['mean_100ep_reward'] > max_mean_reward
-    ):
+          and locals['num_episodes'] >= 500
+          and ( ((locals['num_episodes']-best_reward_episode)%50 ==0) or (locals['mean_100ep_reward'] > max_mean_reward))
+      ):
       if(not os.path.exists(os.path.join(PROJ_DIR,'models/deepq/'))):
         try:
           os.mkdir(os.path.join(PROJ_DIR,'models/'))
@@ -123,18 +126,23 @@ def deepq_callback(locals, globals):
           os.mkdir(os.path.join(PROJ_DIR,'models/deepq/'))
         except Exception as e:
           print(str(e))
+      if locals['mean_100ep_reward'] > max_mean_reward:
+        if(last_filename != ""):
+          os.remove(last_filename)
+          print("delete last model file : %s" % last_filename)
 
-      if(last_filename != ""):
-        os.remove(last_filename)
-        print("delete last model file : %s" % last_filename)
+        max_mean_reward = locals['mean_100ep_reward']
+        best_reward_episode = locals['num_episodes']
+        act = dqfd.ActWrapper(locals['act'])
 
-      max_mean_reward = locals['mean_100ep_reward']
-      act = dqfd.ActWrapper(locals['act'])
-
-      filename = os.path.join(PROJ_DIR,'models/deepq/zergling_%s.pkl' % locals['mean_100ep_reward'])
-      act.save(filename)
-      print("save best mean_100ep_reward model to %s" % filename)
-      last_filename = filename
+        filename = os.path.join(PROJ_DIR,'models/deepq/zergling_%s.pkl' % locals['mean_100ep_reward'])
+        act.save(filename)
+        print("save best mean_100ep_reward model to %s" % filename)
+        last_filename = filename
+      else:
+        act = dqfd.ActWrapper(locals['act'])
+        filename = os.path.join(PROJ_DIR, 'models/deepq/zergling_%s.pkl' % locals['mean_100ep_reward'])
+        act.save(filename)
 
 
 if __name__ == '__main__':
